+++ src/controllers/productsController.js
@@ -1,18 +1,15 @@
 const fs = require('fs');
 const path = require('path');
 
-const productsFilePath = path.join(__dirname, '../db/productsDataBase.json');
+const productsFilePath = path.join(__dirname, '../db/products.json');
 const products = JSON.parse(fs.readFileSync(productsFilePath, 'utf-8'));
 
-const toThousand = n => n.toString().replace(/\B(?=(\d{3})+(?!\d))/g, ".");
 
 const controller = {
-	// Root - Show all products
 	index: (req, res) => {
 		id = req.params.id;
-		return res.render("products", {products: products, id: id})
+		return res.render("index", {products})
 	},
-
 	// Detail - Detail from one product
 	detail: (req, res) => {
 		id = req.params.id;
@@ -45,7 +42,21 @@ const controller = {
 	// Delete - Delete one product from DB
 	destroy : (req, res) => {
 		return res.send("Producto borrado exitosamente")
-	}
+	},
+	shop: (req, res) => {
+        return res.render("shop")
+    },
+	cart: (req, res) => {
+        return res.render("cart")
+    },
+
+    productDetail: (req, res) => {
+        return res.render("productdetail")
+    },
+
+    productForm: (req, res) => {
+        return res.render("productForm")
+    }
 };
 
 module.exports = controller;
