--- src/routes/products.js
@@ -1,28 +1,0 @@
-// ************ Require's ************
-const express = require('express');
-const router = express.Router();
-
-// ************ Controller Require ************
-const productsController = require('../controllers/productsController');
-
-/*** GET ALL PRODUCTS ***/ 
-router.get('/', productsController.index); 
-
-/*** CREATE ONE PRODUCT ***/ 
-router.get('/create/', productsController.create); 
-router.post('/', productsController.store); 
-
-
-/*** GET ONE PRODUCT ***/ 
-router.get('/:id/', productsController.detail); 
-
-/*** EDIT ONE PRODUCT ***/ 
-router.put('/:id/edit', productsController.edit); 
-router.put('/:id', productsController.update); 
-
-
-/*** DELETE ONE PRODUCT***/ 
-router.delete('/:id', productsController.destroy); 
-
-
-module.exports = router;
