+++ src/routes/usersRoutes.js
@@ -1,0 +1,10 @@
+const express = require('express');
+const router = express.Router();
+
+// ************ Controller Require ************
+const usersController = require('../controllers/usersController');
+
+router.get("/login", usersController.login);
+router.get("/register", usersController.register); 
+
+module.exports = router;
