+++ src/views/index.ejs
@@ -88,12 +88,12 @@
                         </div>
                         <div class="product-conteiner">
                             <div class="product-exhibitor-first">
-                                <a href="/productdetail.html"><div class="product-div">
+                                <a href="/productdetail"><div class="product-div">
                                     <img src="/img/crystalweedoil.jpg" alt="Producto">
                                     <p class="product-name">Product 1</p>   
                                     <h5 class="product-price"> $10.00</h5>
                                 </div></a>
-                                <a href="/productdetail.html"><div class="product-div">
+                                <a href="/productdetail"><div class="product-div">
                                     <img src="/img/creamcaramel.jpg" alt="Producto">
                                     <p class="product-name">Product 1</p>   
                                     <h5 class="product-price"> $10.00</h5>
@@ -118,22 +118,22 @@
                                 </div></a>
                             </div> -->
                             <div class="product-exhibitor-third">
-                                <a href="/productdetail.html"><div class="product-div">
+                                <a href="/productdetail"><div class="product-div">
                                     <img src="/img/crystalweedoil.jpg" alt="Producto">
                                     <p class="product-name">Product 1</p>   
                                     <h5 class="product-price"> $10.00</h5>
                                 </div></a>
-                                <a href="/productdetail.html"><div class="product-div">
+                                <a href="/productdetail"><div class="product-div">
                                     <img src="/img/creamcaramel.jpg" alt="Producto">
                                     <p class="product-name">Product 1</p>   
                                     <h5 class="product-price"> $10.00</h5>
                                 </div></a>
-                                <a href="/productdetail.html"><div class="product-div">
+                                <a href="/productdetail"><div class="product-div">
                                     <img src="/img/amnesialemon.jpg" alt="Producto">
                                     <p class="product-name">Product 1</p>   
                                     <h5 class="product-price"> $10.00</h5>
                                 </div></a>
-                                <a href="/productdetail.html"><div class="product-div">
+                                <a href="/productdetail"><div class="product-div">
                                     <img src="/img/noratropical.jpg" alt="Producto">
                                     <p class="product-name">Product 1</p>   
                                     <h5 class="product-price"> $10.00</h5>
