+++ src/views/partials/head.ejs
@@ -6,9 +6,10 @@
     <link rel="stylesheet" href="/css/productForm.css">
     <link rel="stylesheet" href="/css/styleHome.css">
     <link rel="stylesheet" href="/css/LoginRegister.css">
-    <!-- <link rel="stylesheet" href="/css/productdetail.css"> -->
+    <link rel="stylesheet" href="/css/productdetail.css">
     <link rel="stylesheet" href="/css/LoginRegister.css">
     <link rel="stylesheet" href="/css/stylescart.css"> 
+    <link rel="stylesheet" href="/css/shop.css">
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css" rel="stylesheet" />
 <link rel="preconnect" href="https://fonts.googleapis.com">
 <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
