+++ src/views/productForm.ejs
@@ -1,8 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
     <%-include ("./partials/head.ejs") %>
-</head>
 <body>
     <%-include ("./partials/header.ejs") %>
