+++ src/views/shop.ejs
@@ -1,7 +1,7 @@
 <!DOCTYPE html>
 <html lang="en">
     <%-include ("./partials/head.ejs") %>
-    <link rel="stylesheet" href="/css/shop.css">
+    
 
 <body>
